Various additions and bugfixes
Added a AmmoBox that updates the players ammo, fixed a bug with the current UI deleting 15 ammo at a time instead of only taking how many it needs. fixed some logical errors with the max health. Fixed some typos in the naming convention. Added a "potential rewards" display that tells the user the current multiplier for the rewards. Updated the health bar to not go past the max health(100). Reworked the score system to be a tier system based on time, rather than the system be based off of how many seconds are left, it is now a flat multiplier.