updated ammo stuff added some setters..
I don't care how spaghetti this shit is anymore I'm just going with it and this is how we are going to do ammo. I tried several different solutions and all of them didn't work with how the Ai interacts with our ammo from GunScripts, it was a complete nightmare and I could probably find another solution(Like rebuilding everything from the ground up) but I am sick of fucking with this stupid mechanic. Anyways, this is the current solution to ammo. I know it's not ideal, or really the best. But this is what I have currently and it's been 2 days of failed iterations and this spaghetti monster is the result of that.